Add an option to prevent errors in new compilers
I compared with `std::move` and without it. Then without it recorded
better performance than with it so I decided to suppress this warning.
I am not completely understanding the machanism of RVO and the meaning of
this warning. I'm planning to investigate about this later, but this time, I use
this way for a moment.